Move /repositories to root project
