+++ handlers/todo_handler.go
@@ -6,7 +6,7 @@ import (
 	"strconv"
 
 	"go-todos-api/models"
-	repos "go-todos-api/models/repositories"
+	repos "go-todos-api/repositories"
 
 	"github.com/gin-gonic/gin"
 )
